Obtener Límites de los Barrios y Parajes de la República Dominicana

Descarga (si es necesario) y carga los límites de los barrios y parajes de la República Dominicana, según la division correspondiente al id suministrado.

Uso

gd_bparajes(id = "RD_BPARAJES", sf = TRUE)

Argumentos

id

Nombre del archivo de datos en el servidor remoto.

sf

Lógico. Si es FALSE, devuelve un data.frame regular sin la columna de geometría. Por defecto es TRUE.

Valor

Un objeto de la clase sf o un data.frame.

Detalles

Las divisiones disponibles son:

"RD_BPARAJES" (por defecto): Barrios y Parajes de la República Dominicana.

Ejemplos

if (FALSE) { # \dontrun{
  # Cargar el objeto sf completo
  bparajes_sf <- gd_bparajes()

  # Cargar solo la tabla de atributos (sin geometría)
  bparajes_df <- gd_bparajes(sf = FALSE)
} # }