Limpiar nombres de provincias de República Dominicana

Esta función limpia y estandariza los nombres de las provincias de la República Dominicana, utilizando un dataset de alias para mayor robustez en el emparejamiento de nombres.

Uso

gd_clean_prov_name(prov, .tol = 0.25, .on_error = "fail")

Argumentos

prov

Vector de caracteres con nombres de provincias a limpiar

.tol

Nivel de tolerancia numérico para similitud de cadenas. Por defecto 0.25.

.on_error

Método de manejo de errores. Por defecto "fail". Puede ser "fail" para detener la ejecución, "omit" para ignorar nombres no emparejados, o "na" para devolver NA en nombres no emparejados.

Valor

Vector de caracteres con nombres de provincias limpiados

Ejemplos

if (FALSE) { # \dontrun{
  # Uso básico
  provincias_limpias <- gd_clean_prov_name(c("azua", "barahona", "stgo"))
  
  # Con mayor tolerancia
  provincias_limpias <- gd_clean_prov_name(c("azua", "barahona"), .tol = 0.5)
} # }